% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.r
\docType{data}
\name{continuous}
\alias{continuous}
\title{continuous.rda}
\format{
A data frame with variables:
\describe{
  \item{schid}{School ID}
  \item{id}{Student ID}
  \item{sex}{0 = boys; 1 = girls}
  \item{race}{0 = White; 1 = Others}
  \item{pretest}{Pre test scores}
  \item{stdscore}{Standardized scores}
  \item{stdscore}{Standardized scores}
  \item{trt}{Treatment assignment; 0 = control, 1 = treatment}
  \item{Y}{Outcome}
  \item{q1-q20}{20 continuous items }
}
}
\source{
CTA1
}
\usage{
continuous
}
\description{
A data set containing continuous items information.
Example data regenerated from CTA1
}
\examples{
data(continuous)
summary(continuous)
}
\keyword{datasets}
