% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.r
\docType{data}
\name{example1}
\alias{example1}
\title{example1.rda}
\format{
A data frame with variables:
\describe{
  \item{Y}{Outcome}
  \item{trt}{Treatment assignment; 0 = control, 1 = treatment}
  \item{X1}{Continuous covariate 1}
  \item{X2}{Continuous covariate 2}
  \item{q1}{Binary item}
  \item{q2}{Binary item}
  \item{q3}{Binary item}
  \item{q4}{Binary item}
  \item{q5}{Binary item}
  \item{q6}{Binary item}
  \item{q7}{Binary item}
  \item{q8}{Binary item}
  \item{q9}{Binary item}
  \item{q10}{Binary item}
}
}
\usage{
example1
}
\description{
A generated data set containing binary items information.
}
\keyword{datasets}
