% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.r
\docType{data}
\name{example2}
\alias{example2}
\title{example2.rda}
\format{
A data frame with variables:
\describe{
  \item{Y}{Outcome}
  \item{trt}{Treatment assignment; 0 = control, 1 = treatment}
  \item{X1}{Continuous covariate 1}
  \item{X2}{Continuous covariate 2}
  \item{q1}{Continuous item}
  \item{q2}{Continuous item}
  \item{q3}{Continuous item}
  \item{q4}{Continuous item}
  \item{q5}{Continuous item}
  \item{q6}{Continuous item}
  \item{q7}{Continuous item}
  \item{q8}{Continuous item}
  \item{q9}{Continuous item}
  \item{q10}{Continuous item}
}
}
\usage{
example2
}
\description{
A generated data set containing continuous items information.
}
\keyword{datasets}
