% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.r
\docType{data}
\name{example3}
\alias{example3}
\title{example3.rda}
\format{
A data frame with variables:
\describe{
  \item{Y}{Outcome}
  \item{Z}{Treatment assignment; 0 = control, 1 = treatment}
  \item{X1}{Continuous covariate 1}
  \item{X2}{Continuous covariate 2}
  \item{X3}{Continuous covariate 2}
  \item{X4}{Continuous covariate 2}
  \item{V1}{Continuous item}
  \item{V2}{Continuous item}
  \item{V3}{Continuous item}
  \item{V4}{Continuous item}
  \item{V5}{Continuous item}
  \item{V6}{Continuous item}
  \item{V7}{Continuous item}
  \item{V8}{Continuous item}
  \item{V9}{Continuous item}
  \item{V10}{Continuous item}
  \item{V11}{Continuous item}
  \item{V12}{Continuous item}
}
}
\usage{
example3
}
\description{
A generated data set for multidimensional FLPS.
}
\keyword{datasets}
