% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flps-package.r
\docType{package}
\name{flps-package}
\alias{flps-package}
\alias{flps}
\title{Fully latent principal stratification}
\description{
The FLPS package conducts Bayesian analysis for fully latent principal stratification via rstan.
}
\details{
The 'flps' package.
}
\references{
Sales, A. C., & Pane, J. F. (2019). The role of mastery learning in an intelligent tutoring system: Principal stratification on a latent variable. The Annals of Applied Statistics, 13(1), 420-443.
Lee, S., Adam, S., Kang, H.-A., & Whittaker, T. A. (2022). Fully latent principal
stratification: Combining ps with model-based measurement models. In The annual
meeting of the psychometric society (pp. 287–298).
}
\author{
Sooyong Lee \email{sooyongl09@gmail.com}
}
\keyword{package}
