% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.r
\name{flps_latent}
\alias{flps_latent}
\title{Latent factor scores distribution by treatment assignment}
\usage{
flps_latent(object, type = "hist")
}
\arguments{
\item{object}{a flps object}

\item{type}{a character indicating the type of plots}
}
\description{
Latent factor scores distribution by treatment assignment
}
