% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rstan.r
\name{importModel}
\alias{importModel}
\title{Import complied Stan object}
\usage{
importModel(
  type = c("rasch", "irt", "grm", "sem", "lca", "lpa"),
  multilevel = FALSE
)
}
\arguments{
\item{type}{a character indicating the type of FLPS model.}

\item{multilevel}{a logical indicating multilevel Stan model.}
}
\value{
a Stan complied object generated by \code{modelBuilder}
}
\description{
Import complied Stan object
}
