% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rstan.r
\name{loadRstan}
\alias{loadRstan}
\title{Load rstan model}
\usage{
loadRstan(lv_type = "2PL", force.string = F)
}
\arguments{
\item{lv_type}{a character specifying a latent model}

\item{force.string}{a logical indicating forcing the string stan code}
}
\value{
A string for Stan syntax
}
\description{
Load rstan model
}
