% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FLPSdata.r
\name{makeFLPSdata}
\alias{makeFLPSdata}
\title{Convert a matrix to a FLPS data}
\usage{
makeFLPSdata(
  inp_data,
  outcome,
  group,
  covariate,
  lv_model,
  lv_type,
  custom = F,
  ...
)
}
\arguments{
\item{inp_data}{A matrix or a data frame}

\item{outcome}{A character indicating the name of an outcome variable}

\item{group}{A character indicating the name of a treatment/control group variable}

\item{covariate}{A character indicating the names of covariates variables}

\item{lv_model}{A description of the latent variable model, which is similar to the lavaan model syntax.}

\item{lv_type}{A character indicating the type of latent variable models}

\item{custom}{a logical indicating whether the cumsom data is used.}

\item{...}{Additional arguments for latent variable models information (e.g., nclass = 2).}
}
\value{
a S3 class for a corresponding measurement model
}
\description{
Convert a matrix to a FLPS data
}
