% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_make_data.r
\name{makeInpData}
\alias{makeInpData}
\title{Generate a matrix style data for simulation}
\usage{
makeInpData(
  N,
  R2Y,
  R2eta,
  omega,
  tau0,
  tau1,
  linear = T,
  ydist = "n",
  lambda,
  nitem,
  nfac = 1,
  lvmodel,
  fcovmat,
  item.missing,
  misspec,
  cov.res,
  relsize
)
}
\arguments{
\item{N}{a numeric indicating sample size.}

\item{R2Y}{a numeric indicating predictive power of covariates.}

\item{R2eta}{a numeric indicating Predictive power of latent variable}

\item{omega}{a numeric indicating the size of effect of latent factor on
the outcome.}

\item{tau0}{a numeric indicating the size of difference in the outcome
between the treatment and the control.}

\item{tau1}{a numeric indicating the principal effect}

\item{linear}{a logical}

\item{ydist}{a character}

\item{lambda}{a numeric indicating the mean of Worked problems/person.
(extent to which covariates predict eta).}

\item{nitem}{a numeric indicating the number of maximum measurement items
given to students.}

\item{nfac}{a numeric indicating the number of latent factors}

\item{lvmodel}{a character specifying a type of latent variable model.}

\item{fcovmat}{a matrix indicating the variance-covariance matrix of latent
factors when nfac > 1}

\item{item.missing}{a logical to make the measurement item data missing for
the control group. Default is TRUE}

\item{misspec}{a logical to allow cross-loadings across latent factors
when nfac > 1.}

\item{cov.res}{a logical to allow for residual correlations
(only for CFA model).}

\item{relsize}{a numeric indicating the degree to which the latent factor explain the variances of continous items (only for CFA model).}
}
\value{
a list containing all the data related to population values and running FLPS.
}
\description{
\code{\link{makeInpDat}} is a function for generating a data based on
the given information.
}
\examples{
sdat <- makeInpData(
  N = 100,
  R2Y = 0.2,
  omega = 0.2,
  tau0 = 0.13,
  tau1 = -0.06,
  lambda = 10,
  R2eta = 0.5,
  nitem = 10,
  linear = T,
  lvmodel = "sem"
)
}
