% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rstan.r
\name{mkStanModel}
\alias{mkStanModel}
\title{Create Stanmodel class}
\usage{
mkStanModel(lv_type = "2pl")
}
\arguments{
\item{lv_type}{a character specifying a latent model}
}
\value{
A stanmodel class
}
\description{
Create Stanmodel class
}
