% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rstan.r
\name{modelBuilder}
\alias{modelBuilder}
\title{Generate compiled Stan object to facilitate the analysis}
\usage{
modelBuilder(type = "all", multilevel = FALSE)
}
\arguments{
\item{type}{A character string specifying the type of FLPS model.
If not provided, the default is set to \code{"all"}, which compiles all available Stan syntax.}

\item{multilevel}{a logical indicating multilevel Stan model.}
}
\value{
There's no return, but the compiled objects are saved in the package
root directory.
}
\description{
Generate compiled Stan object to facilitate the analysis
}
