% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.r
\name{plot.flps}
\alias{plot.flps}
\title{Plot}
\usage{
\method{plot}{flps}(object, type = NULL, pars = c("tau0", "tau1"), ...)
}
\arguments{
\item{object}{an object of class \code{flps}}

\item{type}{a string for the type of plot}

\item{...}{additional options for future development}
}
\description{
Plot
}
