% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.r
\name{plot.flps}
\alias{plot.flps}
\title{Plot}
\usage{
\method{plot}{flps}(x, type = NULL, pars = c("tau0", "tau1"), ...)
}
\arguments{
\item{x}{an object of class \code{flps}}

\item{type}{a string for the type of plot}

\item{pars}{a character vector indicating the target parameters}

\item{...}{additional options for future development}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further customized
using the \pkg{ggplot2} package.
}
\description{
Plot
}
