% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.r
\name{print.flps}
\alias{print.flps}
\title{Print results}
\usage{
\method{print}{flps}(obj, ...)
}
\arguments{
\item{obj}{an object of class \code{flps}}

\item{...}{additional options for future development}
}
\description{
Print results
}
