% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.r
\name{print.summary.flps}
\alias{print.summary.flps}
\title{Print summary of results}
\usage{
\method{print}{summary.flps}(x, type = "structures", ...)
}
\arguments{
\item{x}{an object of class \code{\link[flps]{flps}}}

\item{type}{a string for the part of FLPS model
\itemize{
  \item \code{structures} : prints the results of structural parts.
  \item \code{measurement} : prints the results of measurement parts.
  \item \code{latent} : prints the information of individual latent scores
  \item \code{raw} : prints the results via the \code{summary} function of \pkg{rstan} package..
}}

\item{...}{additional options for future development}
}
\value{
Summary of FLPS model are printed.
}
\description{
Print summary of results
}
