% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FLPS.r
\name{runFLPS}
\alias{runFLPS}
\title{Conduct fully latent principal stratification}
\usage{
runFLPS(
  inp_data = NULL,
  custom_data = NULL,
  custom_stan = NULL,
  outcome = NULL,
  group = NULL,
  covariate = NULL,
  lv_model = NULL,
  lv_type = NULL,
  stan_options = list(),
  ...
)
}
\arguments{
\item{inp_data}{A matrix or a data frame}

\item{custom_data}{A list. should be provided with custom_stan.}

\item{custom_stan}{A string. should be provided with custom_data.}

\item{outcome}{A character indicating the name of an outcome variable}

\item{group}{A character indicating the name of a treatment/control group variable}

\item{covariate}{A character indicating the names of covariates variables}

\item{lv_model}{A description of the latent variable model, which is similar to the \pkg{lavaan} model syntax.}

\item{lv_type}{A character indicating the type of latent variable models}

\item{stan_options}{A list containing [rstan::stan()] options, using 'name = value'.}

\item{...}{Additional arguments for latent variable models information (e.g., nclass = 2).}
}
\value{
an object of class \code{flps} which contains a  \code{\link{stan}} object.
}
\description{
Conduct fully latent principal stratification
}
\seealso{
[rstan::stan()]
}
\concept{rstan}
