% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.r
\docType{data}
\name{binary}
\alias{binary}
\title{binary.rda}
\format{
A data frame with variables:
\describe{
  \item{schid}{School ID}
  \item{id}{Student ID}
  \item{sex}{0 = boys; 1 = girls}
  \item{race}{0 = White; 1 = Others}
  \item{pretest}{Pre test scores}
  \item{stdscore}{Standardized scores}
  \item{trt}{Treatment assignment; 0 = control, 1 = treatment}
  \item{Y}{Outcome}
  \item{q1}{Binary item}
  \item{q2}{Binary item}
  \item{q3}{Binary item}
  \item{q4}{Binary item}
  \item{q5}{Binary item}
  \item{q6}{Binary item}
  \item{q7}{Binary item}
  \item{q8}{Binary item}
  \item{q9}{Binary item}
  \item{q10}{Binary item}
  \item{q11}{Binary item}
  \item{q12}{Binary item}
  \item{q13}{Binary item}
  \item{q14}{Binary item}
  \item{q15}{Binary item}
  \item{q16}{Binary item}
  \item{q17}{Binary item}
  \item{q18}{Binary item}
  \item{q19}{Binary item}
  \item{q20}{Binary item}
}
}
\source{
CTA1
}
\usage{
binary
}
\description{
A data set containing binary items information.
Example data regenerated from CTA1
}
\examples{
data(binary)
summary(binary)
}
\keyword{datasets}
