% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.r
\docType{data}
\name{continuous}
\alias{continuous}
\title{continuous.rda}
\format{
A data frame with variables:
\describe{
  \item{schid}{School ID}
  \item{id}{Student ID}
  \item{sex}{0 = boys; 1 = girls}
  \item{race}{0 = White; 1 = Others}
  \item{pretest}{Pre test scores}
  \item{stdscore}{Standardized scores}
  \item{cm_sex}{Cluster-mean of sex}
  \item{cm_race}{Cluster-mean of race}
  \item{cm_pretest}{Cluster-mean of Pre test scores}
  \item{cm_stdscore}{Cluster-mean of of Standardized scores}
  \item{trt}{Treatment assignment; 0 = control, 1 = treatment}
  \item{Y}{Outcome}
  \item{q1}{Continuous item}
  \item{q2}{Continuous item}
  \item{q3}{Continuous item}
  \item{q4}{Continuous item}
  \item{q5}{Continuous item}
  \item{q6}{Continuous item}
  \item{q7}{Continuous item}
  \item{q8}{Continuous item}
  \item{q9}{Continuous item}
  \item{q10}{Continuous item}
  \item{q11}{Continuous item}
  \item{q12}{Continuous item}
  \item{q13}{Continuous item}
  \item{q14}{Continuous item}
  \item{q15}{Continuous item}
  \item{q16}{Continuous item}
  \item{q17}{Continuous item}
  \item{q18}{Continuous item}
  \item{q19}{Continuous item}
  \item{q20}{Continuous item}
}
}
\source{
CTA1
}
\usage{
continuous
}
\description{
A data set containing continuous items information.
Example data regenerated from CTA1
}
\examples{
data(continuous)
summary(continuous)
}
\keyword{datasets}
