% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flps-package.r
\docType{package}
\name{FLPS-package}
\alias{FLPS-package}
\alias{flps}
\title{Fully latent principal stratification}
\description{
The FLPS package conducts Bayesian analysis for fully latent principal stratification via rstan.
}
\details{
The 'flps' package.
}
\references{
Sales, A. C., & Pane, J. F. (2019). The role of mastery learning in an intelligent tutoring system: Principal stratification on a latent variable. The Annals of Applied Statistics, 13(1), 420-443.
}
\author{
Sooyong Lee \email{sooyongl09@utexas.edu}
}
\keyword{package}
