% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.r
\name{flps_plot}
\alias{flps_plot}
\title{Make plots related to FLPS models}
\usage{
flps_plot(object, type = "latent")
}
\arguments{
\item{object}{a flps object}

\item{type}{a character indicating the type of plots}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further customized
  using the \pkg{ggplot2} package.
}
\description{
Make plots related to FLPS models
}
