% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rstan.r
\name{importModel}
\alias{importModel}
\title{Import compiled Stan object}
\usage{
importModel(lv_type, multilevel = FALSE, lv_randomeffect = FALSE)
}
\arguments{
\item{lv_type}{a character indicating the type of FLPS model.}

\item{multilevel}{a logical indicating multilevel Stan model.}

\item{lv_randomeffect}{A logical indicating whether to estimate random effects for latent variables.}
}
\value{
a Stan compiled stanmodel object generated by \code{modelBuilder}
}
\description{
Import compiled Stan object
}
