% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rstan.r
\name{importModel}
\alias{importModel}
\title{Import compiled Stan object}
\usage{
importModel(type, multilevel = FALSE)
}
\arguments{
\item{type}{a character indicating the type of FLPS model.}

\item{multilevel}{a logical indicating multilevel Stan model.}
}
\value{
a Stan compiled \code{\link[rstan]{stanmodel}} object generated by \code{modelBuilder}
}
\description{
Import compiled Stan object
}
