% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rstan.r
\name{modelBuilder}
\alias{modelBuilder}
\title{Generate compiled Stan object to facilitate the analysis}
\usage{
modelBuilder(type = "all")
}
\arguments{
\item{type}{a character indicating the type of FLPS model. The default is NULL to compile all available Stan codes ('all' does the same thing).}
}
\value{
There's no return, but the compiled objects are saved in the package root directory.
}
\description{
Generate compiled Stan object to facilitate the analysis
}
