% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.r
\name{print.flps}
\alias{print.flps}
\title{Print results}
\usage{
\method{print}{flps}(x, ...)
}
\arguments{
\item{x}{an object of class \code{flps}}

\item{...}{additional options for future development}
}
\value{
Results of FLPS model are printed via the \pkg{rstan} package.
}
\description{
Print results
}
