% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.r
\name{summary.flps}
\alias{summary.flps}
\title{Summarize the results}
\usage{
\method{summary}{flps}(object, type = "all", ...)
}
\arguments{
\item{object}{an object of class \code{flps}}

\item{type}{a string for the part of FLPS model}

\item{...}{additional options for future development}
}
\value{
Summary of FLPS model are printed via the \pkg{rstan} package.
}
\description{
Summarize the results
}
