% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FLPS.r
\name{summary.flps}
\alias{summary.flps}
\title{Summarize the results}
\usage{
summary.flps(object, type = "all")
}
\arguments{
\item{object}{an object of class \code{flps}}

\item{type}{a string for the part of FLPS model}
}
\description{
Summarize the results
}
